Add two Reader tests to Native
