+++ sconfig/shared/src/test/scala/org/ekrich/config/ConfigFactoryTests.scala
@@ -1,9 +1,24 @@
 package org.ekrich.config
 
+// import java.io.StringReader
+
 import org.junit.Assert._
 import org.junit.Test
 
 class ConfigFactoryTests {
+
+  // uncomment when Scala.js has Reader
+  // val filename = "/test01.properties"
+  // val fileStr =
+  //   """
+  //     |# test01.properties file
+  //     |fromProps.abc=abc
+  //     |fromProps.one=1
+  //     |fromProps.bool=true
+  //     |fromProps.specialChars=hello^^
+  //     """.stripMargin
+  // var test01Reader = new StringReader(fileStr)
+
   @Test def parseString: Unit = {
     val configStr =
       """
@@ -41,4 +56,26 @@ class ConfigFactoryTests {
     val pattern = rconf.getList("core.extends").get(0).unwrapped
     assertEquals(pattern, "default")
   }
+
+  // @Test
+  // def parse(): Unit = {
+  //   val filename = "/test01.properties"
+  //   val config = ConfigFactory.parseReader(
+  //     test01Reader,
+  //     ConfigParseOptions.defaults
+  //       .setSyntaxFromFilename(filename)
+  //   )
+  //   assertEquals("hello^^", config.getString("fromProps.specialChars"))
+  // }
+
+  // @Test
+  // def parseIncorrectFormat(): Unit = {
+  //   val e = assertThrows(
+  //     classOf[ConfigException.Parse],
+  //     () => ConfigFactory.parseReader(test01Reader)
+  //   )
+  //   assertTrue(
+  //     e.getMessage.contains("Expecting end of input or a comma, got '^'")
+  //   )
+  // }
 }
